FIx Account Activate process (missing parameter). Add ModelState checking to Deletion.

+++ nexus-broker-sample/Pages/Account/Delete.cshtml.cs
@@ -27,6 +27,11 @@ namespace Nexus.Samples.Broker.Pages.Account
 
         public async Task<IActionResult> OnPostAsync()
         {
+            if (!ModelState.IsValid)
+            {
+                return Page();
+            }
+
             var deleteResult = await nexusClient.DeleteAccount(AccountCode);
 
             if (!deleteResult.IsSuccess)
